third_party/WebKit/LayoutTests/fast/backgrounds/size/parsing-inherit-expected.txt

This test checks that background-size:inherit properly inherits all values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test("background-size: contain;") is "contain"
PASS test("background-size: cover;") is "cover"
PASS test("background-size: 100 100;") is "auto"
PASS test("background-size: 100px 100px;") is "100px 100px"
PASS test("background-size: auto 50px;") is "auto 50px"
PASS test("background-size: 50px auto;") is "50px"
PASS test("background-size: auto auto;") is "auto"
PASS test("background-size: 30% 20%;") is "30% 20%"
PASS test("background-size: 4em auto;") is "64px"
PASS test("background-size: 5em ;") is "80px"
PASS test("-webkit-background-size: 5em ;") is "80px 80px"
PASS test("background-size: 100px 100px 100px;") is "auto"
PASS test("background-size: coconut;") is "auto"
PASS successfullyParsed is true

TEST COMPLETE
